Moved method logic handling to lib directory
https://github.com/vercel/platforms/pull/25#discussion_r791263553
